GMT_DOS-CLIENTS_DCS RENAME MAIN & README
`main` is renamed `mount`
