+++ clients/dcs/Cargo.toml
@@ -16,7 +16,7 @@ env_logger.workspace = true
 gmt_dos-actors.workspace = true
 gmt_dos-clients = { workspace = true, features = ["nalgebra"] }
 gmt_dos-clients_arrow.workspace = true
-gmt_dos-clients_fem.workspace = true
+gmt_dos-clients_fem = { workspace = true, features = ["serde"] }
 gmt_dos-clients_io.workspace = true
 gmt_dos-clients_mount.workspace = true
 gmt_dos-clients_scope = { workspace = true, features = ["server"] }
